backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.12.5
pydantic-settings==2.12.0
qdrant-client==1.12.2
sqlalchemy==1.4.54
psycopg2-binary==2.9.11

google-generativeai==0.8.4
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
python-dateutil==2.9.0
aiofiles==24.1.0
cryptography==44.0.2
slowapi==0.1.9